requirements.txt
mediapipe==0.10.2
Pillow==8.3.1
fastapi==0.100.0
uvicorn==0.23.1
requests==2.31.0